chore: move to root directory
Since the exercise are built on top of each other we are in no need of having different folders for them. For this reason this commit moves the whole source code to the root directory.
Also this commit inlines the *Section components, since there are not used multiple teams and dont provide a meaningful seperation, as similar named components already exist.